Flask Removed, Jinja2 added and argparse added

+++ pub/main.py
@@ -1,28 +1,37 @@
-from flask import Flask, render_template
 import yaml
 import pdfkit
-import pdfcrowd
 
-app = Flask(__name__)
+import jinja2
+from jinja2 import Template
 
-with app.app_context():
-    stream = open("index.yaml", "r")
-    docs = yaml.load(stream)
-    for doc in docs:
-        filename=docs[doc]["name"]+".html"
-        filename1=docs[doc]["name"]+".pdf"
-        f= open(filename,"w")
-        f.truncate()
-        stream1 = open(docs[doc]["file_name"], "r")
-        items = yaml.load(stream1)
-        f.write(render_template("document.html", items=items))
-        f.close()
-        options = {
-            'page-size': 'A4',
-            'dpi': 250
-        }
-        pdfkit.from_url("http://localhost:8888/restdoc/pub/"+filename, filename1, options=options)
-    exit()
+import argparse
 
-if __name__ == "__main__":
-    app.run()
+parser = argparse.ArgumentParser()
+parser.add_argument("Yaml_File", help="Path of the Yaml File")
+parser.add_argument("out_dir", help="Output Directory")
+parser.add_argument("out_type", choices=["PDF", "HTML", "BOTH"], help="Output Type - PDF, HTML or BOTH")
+args = parser.parse_args()
+
+stream = open(args.Yaml_File, "r")
+items = yaml.load(stream)
+
+with open('templates/document.tmpl') as t:
+    tmpl = Template(t.read())
+
+if args.out_type == "PDF" or args.out_type == "BOTH" :
+    filename=args.out_dir+items["name"]+".pdf"
+    options = {
+        'page-size': 'A4',
+        'dpi': 250
+    }
+    pdfkit.from_string(tmpl.render(items=items), filename, options=options)
+
+if args.out_type == "HTML" or args.out_type == "BOTH":
+    filename=args.out_dir+items["name"]+".html"
+
+    f = open(filename,"w")
+    f.truncate()
+
+    f.write(tmpl.render(items=items))
+    f.close()
+
